기안등록 성공.............

+++ ERgateProject/src/main/java/com/kh/ergate/sign/model/vo/SignAttachment.java
@@ -13,7 +13,7 @@ import lombok.ToString;
 @ToString
 public class SignAttachment {
 	private int fileNo; //파일번호
-	private int refDocNo; //문서번호
+	private long refDocNo; //문서번호
 	private String filePath; //파일경로
 	private String originName; //파일원본명
 	private String changeName; //파읿수정명
